requirements.txt
# SFT - Supervised Fine-Tuning for LLM
# 基于 HuggingFace TRL + LoRA 的大语言模型微调项目

# 核心依赖
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.8.0

# 量化支持 (可选，用于 QLoRA)
bitsandbytes>=0.41.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0

# 可视化与日志
tensorboard>=2.15.0

# 开发依赖
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
